@@ -3,18 +3,22 @@ version: 0.6.2+4
 homepage: https://github.com/SunnyApp/worker_service
 description: A flutter package that works for native/web, and allows the creation of workers that offload work into a potentially parallel space
 environment: 
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies: 
   async: ^2.5.0
   equatable: ^2.0.0
   isolate: ^2.0.3
   js: ^0.6.3
   logging: ^1.0.0
-  logging_config: ^0.5.7+1
+  logging_config: ^0.5.7+2
   meta: ^1.3.0
   service_worker: ^0.2.4
-  sunny_dart: ^0.6.0+4
-flutter: 
+  sunny_dart: ^0.8.0+6
+  collection: ^1.15.0-nullsafety.4
+dependency_overrides:
+  sunny_dart: ^0.8.0+6
+
+flutter:
   assets: 
     - bootstrap_dev.js
     - bootstrap_prod.js
