@@ -4,7 +4,7 @@ import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 
 class DecodedMessage {
-  final int messageCode;
+  final int? messageCode;
   final dynamic payload;
 
   DecodedMessage(this.messageCode, [this.payload]);
@@ -14,9 +14,9 @@ class DecodedMessage {
     var l = raw as List;
     assert(l.length == 1 || l.length == 3,
         "Payload must either have 1 or 3 elements but got ${l.length}");
-    final messageCode = raw[0] as int;
+    final messageCode = raw[0] as int?;
     if (l.length > 1) {
-      final contentCode = raw[1] as int;
+      final contentCode = raw[1] as int?;
       final content = raw[2];
       return DecodedMessage(messageCode, env.decode(contentCode, content));
     } else {
@@ -29,7 +29,7 @@ class DecodedMessage {
     var l = raw as List;
     assert(l.length == 1 || l.length == 3,
         "Payload must either have 1 or 3 elements but got ${l.length}");
-    final messageCode = raw[0] as int;
+    final messageCode = raw[0] as int?;
     if (l.length > 1) {
       final content = raw[2];
       return DecodedMessage(messageCode, content);
@@ -48,14 +48,14 @@ class MessageEnvelope {
 }
 
 abstract class PayloadHandler {
-  dynamic decode(int contentType, dynamic content);
+  dynamic decode(int? contentType, dynamic content);
 
   Payload encode(Payload payload);
 
   factory PayloadHandler.of(
-          {int contentType,
-          @required PayloadDecoder decode,
-          @required PayloadEncoder encode}) =>
+          {int? contentType,
+          required PayloadDecoder decode,
+          required PayloadEncoder encode}) =>
       _PayloadHandler(decode, encode, contentType);
 
   static const defaults = _PayloadHandler(_decodeJson, _encodeJson, null);
@@ -85,7 +85,7 @@ dynamic _decodeJson(int type, inbound) {
 }
 
 class _PayloadHandler implements PayloadHandler {
-  final int contentType;
+  final int? contentType;
   final PayloadDecoder _decode;
   final PayloadEncoder _encode;
 
@@ -95,7 +95,7 @@ class _PayloadHandler implements PayloadHandler {
   Payload encode(Payload env) => _encode(env);
 
   @override
-  dynamic decode(int contentType, content) {
+  dynamic decode(int? contentType, content) {
     return _decode(contentType, content);
   }
 }
@@ -108,13 +108,13 @@ class Payload {
   final int header;
   final dynamic data;
 
-  const Payload(int header, this.data) : header = header ?? kraw;
+  const Payload(int? header, this.data) : header = header ?? kraw;
 
   const Payload.raw(this.data) : header = kraw;
 }
 
 typedef PayloadEncoder = Payload Function(Payload payload);
-typedef PayloadDecoder = dynamic Function(int contentType, dynamic payload);
+typedef PayloadDecoder = dynamic Function(int? contentType, dynamic payload);
 
 dynamic rawPayloadEncoder(MessageEnvelope env) {
   return env.payload.data;
