@@ -8,7 +8,7 @@ import 'package:worker_service/work/platform/grunt_platform_webworker.dart';
 
 
 Future<DuplexChannel> createGruntChannel(GruntFactory fn,
-    {@required bool isProduction}) async {
+    {required bool isProduction}) async {
 
   var fullPath = "";
   final package = fn.package;
